+++ R/lmmixed.R
@@ -18,6 +18,7 @@
 #'   intercept variable, e.g. "var1", (automatically convered to "(1 | var1)");
 #'   or, (2) the full \code{lme4} specification, e.g. "(var1 | var2)". Note
 #'   parenthesis MUST be included in (2)2 but NOT included in (1).
+#' @param ... Other arguments to pass to \code{lme4::\link[lme4]{lmer}}.
 #' @return A list of multivariable \code{lme4::\link[lme4]{lmer}} fitted model
 #'   outputs. Output is of class \code{lmerMod}.
 #'
@@ -37,9 +38,9 @@
 #'   lmmixed(dependent, explanatory, random_effect) %>%
 #' 	 fit2df(estimate_suffix=" (multilevel")
 
-lmmixed <- function(.data, dependent, explanatory, random_effect){
+lmmixed <- function(.data, dependent, explanatory, random_effect, ...){
 	# If single term random effect, default to random intercept model
 	if(!grepl("\\|", random_effect)) random_effect = paste0("(1 | ", random_effect, ")")
   lme4::lmer(paste0(dependent, "~", paste(explanatory, collapse="+"), " + ", random_effect),
-             data=.data)
+             data = .data, ...)
 }