+++ R/lmuni.R
@@ -9,6 +9,7 @@
 #' @param .data Dataframe.
 #' @param dependent Character vector of length 1, name of depdendent variable (must be continuous vector).
 #' @param explanatory Character vector of any length: name(s) of explanatory variables.
+#' @param ... Other arguments to pass to \code{\link[stats]{lm}}.
 #' @return A list of multivariable \code{\link[stats]{lm}} fitted model outputs.
 #'   Output is of class \code{lmlist}.
 #'
@@ -27,10 +28,12 @@
 #'   lmuni(dependent, explanatory) %>%
 #'   fit2df()
 #'
-lmuni <- function(.data, dependent, explanatory){
+lmuni <- function(.data, dependent, explanatory, ...){
   result <- list()
   for (i in 1:length(explanatory)){
-    result[[i]] <- lm(paste(dependent, "~", explanatory[i]), data=.data)
+    result[[i]] <- ff_eval(
+    	lm(paste(dependent, "~", explanatory[i]), data = .data, ...)
+    )
   }
   class(result) = "lmlist"
   return(result)